Switch from `config.toml` to `build.rs`
The primary purpose of the per-package `config.toml` files was to set
the default target and required subsystem link arg for drivers. The
problem with its use is that it is a project-level configuration for
cargo, and not a package-level configuration.

Use `build.rs` instead, which will be executed before each package is
built. This allows using `cargo build --target <target>` from the
workspace root and having the correct subsystem set on each package.

Ref: https://doc.rust-lang.org/cargo/reference/build-scripts.html
Signed-off-by: Tim Crawford <tcrawford@system76.com>